requirements.txt
streamlit
pdfplumber
pytesseract
langchain-core
langchain
langchain-community
langchain-huggingface
langchain-pinecone
langchain-text-splitters
numpy
pillow
python-dotenv
google-generativeai
PyMuPDF
python-docx
chromadb
huggingface-hub
sentence-transformers
rank-bm25
regex
fastapi
uvicorn
python-multipart
typing_extensions
pinecone-client